app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
minio
python-multipart
python-dotenv
